services/ephemeris/requirements.txt
ephem==4.1.3
grpcio==1.43.0
protobuf==3.19.1